Added support of primtive types
