+++ Dataverse.Browser/Requests/Converters/WebApiRequestConverter.CRUD.cs
@@ -23,169 +23,10 @@ using static System.Windows.Forms.VisualStyles.VisualStyleElement.Tab;
 
 namespace Dataverse.Browser.Requests.Converter
 {
-    internal class WebApiRequestConverter
+    internal partial class WebApiRequestConverter
 
     {
-        private DataverseContext Context { get; }
-
-        public WebApiRequestConverter(DataverseContext context)
-        {
-
-            this.Context = context ?? throw new ArgumentNullException(nameof(context));
-
-        }
-
-        internal InterceptedWebApiRequest ConvertUnknowRequestToOrganizationRequest(IRequest request)
-        {
-            var url = new Uri(request.Url);
-            var localPathWithQuery = url.LocalPath + url.Query;
-            if (!localPathWithQuery.StartsWith("/api/data/v9."))
-                return null;
-            SimpleHttpRequest simplifiedRequest;
-            try
-            {
-                simplifiedRequest = new SimpleHttpRequest(request, localPathWithQuery);
-            }
-            catch (Exception ex)
-            {
-                return new InterceptedWebApiRequest()
-                {
-                    SimpleHttpRequest = new SimpleHttpRequest() { LocalPathWithQuery = localPathWithQuery, Method = request.Method },
-                    ConvertFailureMessage = ex.Message,
-                    ExecuteException = ex
-                };
-            }
-            return ConvertDataApiSimplifiedRequestToOrganizationRequest(simplifiedRequest);
-        }
-
-        private InterceptedWebApiRequest ConvertUnknowSimplifiedRequestToOrganizationRequest(SimpleHttpRequest request)
-        {
-            if (!request.LocalPathWithQuery.StartsWith("/api/data/v9."))
-                return null;
-            return ConvertDataApiSimplifiedRequestToOrganizationRequest(request);
-        }
-
-        private InterceptedWebApiRequest ConvertDataApiSimplifiedRequestToOrganizationRequest(SimpleHttpRequest request)
-        {
-            InterceptedWebApiRequest webApiRequest = new InterceptedWebApiRequest()
-            {
-                SimpleHttpRequest = request
-            };
-            ODataUriParser parser;
-            ODataPath path;
-            try
-            {
-                parser = new ODataUriParser(this.Context.Model, new Uri(request.LocalPathWithQuery.Substring(15), UriKind.Relative));
-                path = parser.ParsePath();
-            }
-            catch (Exception ex)
-            {
-                webApiRequest.ConvertFailureMessage = "Unable to parse: " + ex.Message;
-                return webApiRequest;
-            }
-            try
-            {
-                switch (request.Method)
-                {
-                    case "POST":
-                        if (path.Count != 1)
-                        {
-                            throw new NotImplementedException("POST is not implemented for: " + path.Count + " segments");
-                        }
-                        if (path.FirstSegment.EdmType?.TypeKind == EdmTypeKind.Collection)
-                        {
-                            ConvertToCreateUpdateRequest(webApiRequest, path);
-                        }
-                        else if (path.FirstSegment.Identifier == "$batch")
-                        {
-                            if (request.OriginRequest == null)
-                            {
-                                throw new NotSupportedException("batch requests embedded in another batch request are not supported!");
-                            }
-                            ConvertToExecuteMultipleRequest(webApiRequest);
-                        }
-                        else if (path.FirstSegment.EdmType == null && path.FirstSegment is OperationImportSegment operationImport)
-                        {
-                            string identifier = path.FirstSegment.Identifier;
-                            var operation = this.Context.Model.FindDeclaredOperationImports(identifier).Single();
-                            if (operation.IsActionImport())
-                            {
-                                ConvertToAction((IEdmOperation) operation, webApiRequest);
-                            }
-                            else
-                            {
-                                throw new NotImplementedException("Non action operations are not implemented");
-                            }
-                        }
-                        else
-                        {
-                            throw new NotImplementedException("POST is not implemented for: " + path.FirstSegment.EdmType?.TypeKind);
-                        }
-                        break;
-                    case "PATCH":
-                        if (path.Count != 2)
-                        {
-                            throw new NotImplementedException("PATCH is not implemented for: " + path.Count + " segments");
-                        }
-                        if (path.FirstSegment.EdmType?.TypeKind != EdmTypeKind.Collection)
-                        {
-                            throw new NotImplementedException("PATCH is not implemented for: " + path.FirstSegment.EdmType?.TypeKind);
-                        }
-                        ConvertToCreateUpdateRequest(webApiRequest, path);
-                        break;
-                    case "GET":
-                        switch (path.Count)
-                        {
-                            case 1:
-                                throw new NotImplementedException("Retrievemultiple are not implemented");
-                            case 2:
-                                ConvertToRetrieveRequest(webApiRequest, parser, path);
-                                break;
-                            default:
-                                throw new NotSupportedException("Unexpected number of segments:" + path.Count);
-                        }
-                        break;
-                    case "DELETE":
-                        if (path.Count != 2)
-                        {
-                            throw new NotSupportedException("Unexpected number of segments:" + path.Count);
-                        }
-                        ConvertToDeleteRequest(webApiRequest, path);
-                        break;
-                    default:
-                        webApiRequest.ConvertFailureMessage = "method not implemented";
-                        break;
-                }
-            }
-            catch (Exception ex)
-            {
-                webApiRequest.ConvertFailureMessage = ex.Message;
-            }
-            return webApiRequest;
-        }
-
-        private void ConvertToAction(IEdmOperation operation, InterceptedWebApiRequest webApiRequest)
-        {
-            OrganizationRequest request = new OrganizationRequest(operation.Name);
-            using (JsonDocument json = JsonDocument.Parse(webApiRequest.SimpleHttpRequest.Body))
-            {
-                foreach (var node in json.RootElement.EnumerateObject())
-                {
-                    string key = node.Name;
-                    var parameter = operation.FindParameter(key);
-                    if (parameter == null)
-                    {
-                        throw new NotSupportedException($"parameter {key} not found!");
-                    }
-                    request[key] = ConvertValueToAttribute(node.Value, parameter.Type);
-                }
-            }
-        }
-
-        private object ConvertValueToAttribute(JsonElement value, IEdmTypeReference type)
-        {
-            throw new NotImplementedException();
-        }
+       
 
         private void ConvertToDeleteRequest(InterceptedWebApiRequest webApiRequest, ODataPath path)
         {
@@ -274,107 +115,7 @@ namespace Dataverse.Browser.Requests.Converter
             webApiRequest.ConvertedRequest = ConvertToCreateUpdateRequest(keySegment, webApiRequest, entity.LogicalName);
         }
 
-        private OrganizationRequest ConvertToExecuteMultipleRequest(InterceptedWebApiRequest webApiRequest)
-        {
-            var originRequest = webApiRequest.SimpleHttpRequest.OriginRequest;
-            string contentType = originRequest.Headers["Content-Type"];
-            if (!contentType.StartsWith("multipart/mixed;"))
-            {
-                throw new NotImplementedException("ContentType " + contentType + " is not supported for batch requests");
-            }
-
-            ExecuteMultipleRequest executeMultipleRequest = new ExecuteMultipleRequest()
-            {
-                Requests = new OrganizationRequestCollection()
-            };
-
-            MemoryStream dataStream = AddMissingLF(originRequest);
-            using (var content = new StreamContent(dataStream))
-            {
-                //TODO support des changesets
-                //TODO support des continue on error
-                content.Headers.ContentType = MediaTypeHeaderValue.Parse(contentType);
-
-                MultipartMemoryStreamProvider provider = content.ReadAsMultipartAsync().Result;
-
-                //TODO changesets
-                foreach (var httpContent in provider.Contents)
-                {
-                    var data = httpContent.ReadAsByteArrayAsync().Result;
-                    var innerRequest = CreateSimplifiedRequestFromMimeMessage(data);
-                    var convertedRequest = this.ConvertUnknowSimplifiedRequestToOrganizationRequest(innerRequest);
-                    if (convertedRequest == null)
-                    {
-                        throw new NotSupportedException("Only web api requests are supported!");
-                    }
-                    else
-                    if (convertedRequest.ConvertedRequest != null)
-                    {
-                        executeMultipleRequest.Requests.Add(convertedRequest.ConvertedRequest);
-                    }
-                    else
-                    {
-                        throw new NotSupportedException("One inner request could not be converted:" + convertedRequest.ConvertFailureMessage);
-                    }
-                }
-
-            }
-            return executeMultipleRequest;
-        }
-
-        private SimpleHttpRequest CreateSimplifiedRequestFromMimeMessage(byte[] data)
-        {
-            var request = new SimpleHttpRequest();
-            int index = Array.FindIndex(data, b => b == (byte)'\r');
-            if (index == -1)
-            {
-                throw new NotSupportedException("Unable to parse data, no \\r found!");
-            }
-            string firstLine = Encoding.UTF8.GetString(data, 0, index);
-            if (!firstLine.StartsWith("GET"))
-            {
-                throw new ApplicationException("Unable to parse first line: " + firstLine);
-
-            }
-            string url = firstLine.Substring(4);
-            if (url.EndsWith("HTTP/1.1"))
-            {
-                url = url.Substring(0, url.Length - 8);
-            }
-            request.Method = "GET";
-            request.LocalPathWithQuery = url;
-            //TODO : body and headers
-            return request;
-        }
-
-        private static MemoryStream AddMissingLF(IRequest request)
-        {
-            // Les requêtes batch de CRM contiennent uniquement des LF en séparateurs de lignes et pas de CR
-            var data = request.PostData.Elements.FirstOrDefault().Bytes;
-            MemoryStream dataStream = new MemoryStream();
-            bool previousIsCr = false;
-            for (int i = 0; i < data.Length; i++)
-            {
-                var value = data[i];
-                if (value == '\r')
-                {
-                    previousIsCr = true;
-                }
-                else
-                {
-                    if (value == '\n' && !previousIsCr)
-                    {
-                        dataStream.WriteByte((byte)'\r');
-                    }
-                    previousIsCr = false;
-                }
-                dataStream.WriteByte(value);
-            }
-            dataStream.Seek(0, SeekOrigin.Begin);
-            return dataStream;
-        }
-
-        private OrganizationRequest ConvertToCreateUpdateRequest(KeySegment keySegment, InterceptedWebApiRequest webApiRequest, string entityLogicalName)
+       private OrganizationRequest ConvertToCreateUpdateRequest(KeySegment keySegment, InterceptedWebApiRequest webApiRequest, string entityLogicalName)
         {
             string body = webApiRequest.SimpleHttpRequest.Body ?? throw new NotSupportedException("A body was expected!");
             webApiRequest.SimpleHttpRequest.Body = body;
